DFE-XXXX - pushed main existing admin dashboard logic into AdminDashboardPublicationsTab component
